+++ EVENTS.md
@@ -113,27 +113,29 @@ type SocialEventPayload struct {
 
 ### CheckpointEventPayload (Historical Snapshot)
 
-Checkpoints are multi-party attested snapshots of historical state. They anchor restart counts and uptime from before proper event tracking began.
+Checkpoints are multi-party attested snapshots of historical state. They anchor restart counts and uptime from before proper event tracking began. Checkpoints are created through MQTT-based consensus where naras propose and vote on values.
 
 ```go
 type CheckpointEventPayload struct {
     Subject     string   // Who this checkpoint is about
+    SubjectID   string   // Nara ID (for indexing)
     AsOfTime    int64    // Unix timestamp (seconds) when snapshot was taken
     FirstSeen   int64    // When network first saw this nara
     Restarts    int64    // Historical restart count at checkpoint time
     TotalUptime int64    // Total verified online seconds at checkpoint time
     Importance  int      // Always Critical (3) - never pruned
 
-    // Multi-party attestation
-    Attesters   []string // Nara names who attest to this data
-    Signatures  []string // Base64 Ed25519 signatures from attesters
+    // Community consensus - voters who participated in checkpoint creation
+    VoterIDs    []string // Nara IDs who voted for these values
+    Signatures  []string // Base64 Ed25519 signatures (each verifies the values)
 }
 ```
 
 **Key properties:**
 - **Never pruned**: Checkpoints are critical events that survive all pruning
-- **Multi-signed**: Requires multiple high-uptime naras to attest
+- **Multi-signed**: Requires minimum 2 voters (outside proposer) to reach consensus
 - **Historical anchor**: Allows deriving restart count as `checkpoint.Restarts + count(new restarts)`
+- **MQTT consensus**: Created via proposal/vote flow over MQTT topics
 
 ## Personality Filtering
 
@@ -258,26 +260,29 @@ Wait for completion signal...
 
 ## Checkpoint Events (Historical Snapshots)
 
-Checkpoint events capture historical state with multi-party attestation:
+Checkpoint events capture historical state with multi-party consensus:
 
-1. High-uptime naras agree on historical data (restart count, uptime, first-seen time)
-2. They sign a checkpoint event containing this snapshot
-3. Deriving current state: `checkpoint.value + count(events after checkpoint)`
+1. Each nara proposes a checkpoint about itself every 24 hours via MQTT
+2. Other naras vote (approve or reject with their own values)
+3. Consensus is reached via two-round voting (trimmed mean for outliers)
+4. Deriving current state: `checkpoint.value + count(events after checkpoint)`
 
-**Feature flags:**
-- `USE_CHECKPOINTS=true`: Enable reading checkpoint events
-- `USE_CHECKPOINT_CREATION=true`: Participate in creating checkpoints
-- `MIN_CHECKPOINT_ATTESTERS`: Minimum signatures required (default: 2)
-
-**Migration path:**
+**Consensus Flow:**
 ```
-Phase 1: Backfill events spread historical knowledge
+Round 1: Nara proposes {restarts, uptime, first_seen} about itself
     ↓
-Phase 2: High-uptime naras create checkpoints from backfill data
+Voters respond: APPROVE (sign proposal) or REJECT (sign their values)
     ↓
-Phase 3: Checkpoints become the permanent historical anchor
+If majority agrees → checkpoint finalized
+If not → Round 2 with trimmed mean values
     ↓
-Future: Restart count = checkpoint.Restarts + unique StartTimes after checkpoint
+Round 2: Final vote, then give up if no consensus
 ```
+
+**Key properties:**
+- Minimum 2 voters required (outside proposer, so 3+ total signatures)
+- Each signature is for specific values (verifiable)
+- 24-hour cadence per nara
+- Never pruned, always synced
 
 This allows bounded storage while preserving accurate historical data.
