+++ checkpoint_types.go
@@ -5,6 +5,7 @@ import (
 	"crypto/sha256"
 	"encoding/base64"
 	"fmt"
+	"sort"
 	"time"
 )
 
@@ -207,4 +208,222 @@ func (l *SyncLedger) GetCheckpoint(subject string) *CheckpointEventPayload {
 		}
 	}
 	return latest
+}
+
+// DeriveRestartCount derives the total restart count for a subject from checkpoints + events
+// Priority order:
+//  1. If checkpoint exists: checkpoint.Restarts + count(unique StartTimes after checkpoint)
+//  2. If backfill exists: backfill.RestartNum + count(unique StartTimes after backfill timestamp)
+//  3. Otherwise: count(unique StartTimes from all restart events)
+func (l *SyncLedger) DeriveRestartCount(subject string) int64 {
+	l.mu.RLock()
+	defer l.mu.RUnlock()
+
+	// Check for checkpoint first (highest priority)
+	var checkpoint *CheckpointEventPayload
+	var checkpointAsOfTime int64 = 0
+	for _, e := range l.Events {
+		if e.Service == ServiceCheckpoint && e.Checkpoint != nil && e.Checkpoint.Subject == subject {
+			if e.Checkpoint.AsOfTime > checkpointAsOfTime {
+				checkpoint = e.Checkpoint
+				checkpointAsOfTime = e.Checkpoint.AsOfTime
+			}
+		}
+	}
+
+	if checkpoint != nil {
+		// Count unique start times after checkpoint
+		newRestarts := l.countUniqueStartTimesAfterLocked(subject, checkpoint.AsOfTime)
+		return checkpoint.Restarts + newRestarts
+	}
+
+	// Check for backfill event (second priority)
+	var backfill *ObservationEventPayload
+	for _, e := range l.Events {
+		if e.Service == ServiceObservation && e.Observation != nil &&
+			e.Observation.Subject == subject && e.Observation.IsBackfill {
+			backfill = e.Observation
+			break
+		}
+	}
+
+	if backfill != nil {
+		// Count unique start times that are DIFFERENT from the backfill's StartTime
+		// New restarts are identified by having a different StartTime than what backfill captured
+		newRestarts := l.countUniqueStartTimesExcludingLocked(subject, backfill.StartTime)
+		return backfill.RestartNum + newRestarts
+	}
+
+	// No checkpoint or backfill - just count all unique start times
+	return l.countUniqueStartTimesAfterLocked(subject, 0)
+}
+
+// countUniqueStartTimesAfterLocked counts unique StartTime values for restart events after a given time
+// Caller must hold the read lock
+func (l *SyncLedger) countUniqueStartTimesAfterLocked(subject string, afterTime int64) int64 {
+	startTimes := make(map[int64]bool)
+
+	for _, e := range l.Events {
+		if e.Service != ServiceObservation || e.Observation == nil {
+			continue
+		}
+		if e.Observation.Subject != subject {
+			continue
+		}
+		if e.Observation.Type != "restart" {
+			continue
+		}
+		// Skip backfill events when counting new restarts
+		if e.Observation.IsBackfill {
+			continue
+		}
+		// Only count restart events whose StartTime is after the cutoff
+		// This means the restart happened AFTER the checkpoint/reference point
+		if afterTime > 0 && e.Observation.StartTime <= afterTime {
+			continue
+		}
+		startTimes[e.Observation.StartTime] = true
+	}
+
+	return int64(len(startTimes))
+}
+
+// countUniqueStartTimesExcludingLocked counts unique StartTime values excluding a specific one
+// Used for backfill: count all unique start times except the one the backfill already captured
+// Caller must hold the read lock
+func (l *SyncLedger) countUniqueStartTimesExcludingLocked(subject string, excludeStartTime int64) int64 {
+	startTimes := make(map[int64]bool)
+
+	for _, e := range l.Events {
+		if e.Service != ServiceObservation || e.Observation == nil {
+			continue
+		}
+		if e.Observation.Subject != subject {
+			continue
+		}
+		if e.Observation.Type != "restart" {
+			continue
+		}
+		// Skip backfill events when counting new restarts
+		if e.Observation.IsBackfill {
+			continue
+		}
+		// Skip the specific StartTime that the backfill already counted
+		if e.Observation.StartTime == excludeStartTime {
+			continue
+		}
+		startTimes[e.Observation.StartTime] = true
+	}
+
+	return int64(len(startTimes))
+}
+
+// DeriveTotalUptime derives the total uptime for a subject in seconds
+// Priority: checkpoint baseline + uptime from status change events after checkpoint
+//
+// The calculation:
+//  1. Start with checkpoint.TotalUptime (or 0 if no checkpoint)
+//  2. Add up online periods from status-change events after checkpoint
+func (l *SyncLedger) DeriveTotalUptime(subject string) int64 {
+	l.mu.RLock()
+	defer l.mu.RUnlock()
+
+	// Check for checkpoint first
+	var checkpoint *CheckpointEventPayload
+	var checkpointAsOfTime int64 = 0
+	for _, e := range l.Events {
+		if e.Service == ServiceCheckpoint && e.Checkpoint != nil && e.Checkpoint.Subject == subject {
+			if e.Checkpoint.AsOfTime > checkpointAsOfTime {
+				checkpoint = e.Checkpoint
+				checkpointAsOfTime = e.Checkpoint.AsOfTime
+			}
+		}
+	}
+
+	baseUptime := int64(0)
+	afterTime := int64(0)
+	if checkpoint != nil {
+		baseUptime = checkpoint.TotalUptime
+		afterTime = checkpoint.AsOfTime
+	}
+
+	// Collect status change events after checkpoint
+	type statusEvent struct {
+		timestamp int64
+		state     string
+	}
+	var statusEvents []statusEvent
+
+	for _, e := range l.Events {
+		if e.Service != ServiceObservation || e.Observation == nil {
+			continue
+		}
+		if e.Observation.Subject != subject {
+			continue
+		}
+		if e.Observation.Type != "status-change" {
+			continue
+		}
+		eventTimeSec := e.Timestamp / 1e9
+		if afterTime > 0 && eventTimeSec <= afterTime {
+			continue
+		}
+		statusEvents = append(statusEvents, statusEvent{
+			timestamp: eventTimeSec,
+			state:     e.Observation.OnlineState,
+		})
+	}
+
+	// Sort by timestamp
+	sort.Slice(statusEvents, func(i, j int) bool {
+		return statusEvents[i].timestamp < statusEvents[j].timestamp
+	})
+
+	// Calculate online periods
+	var onlineStart int64 = 0
+	for _, se := range statusEvents {
+		if se.state == "ONLINE" && onlineStart == 0 {
+			onlineStart = se.timestamp
+		} else if (se.state == "OFFLINE" || se.state == "MISSING") && onlineStart > 0 {
+			baseUptime += se.timestamp - onlineStart
+			onlineStart = 0
+		}
+	}
+
+	// If still online, count up to now
+	if onlineStart > 0 {
+		baseUptime += time.Now().Unix() - onlineStart
+	}
+
+	return baseUptime
+}
+
+// GetFirstSeenFromEvents returns the earliest known StartTime for a subject from events
+func (l *SyncLedger) GetFirstSeenFromEvents(subject string) int64 {
+	l.mu.RLock()
+	defer l.mu.RUnlock()
+
+	var earliest int64 = 0
+
+	for _, e := range l.Events {
+		// Check observation events (including backfill)
+		if e.Service == ServiceObservation && e.Observation != nil {
+			if e.Observation.Subject == subject && e.Observation.StartTime > 0 {
+				if earliest == 0 || e.Observation.StartTime < earliest {
+					earliest = e.Observation.StartTime
+				}
+			}
+		}
+
+		// Check existing checkpoints
+		if e.Service == ServiceCheckpoint && e.Checkpoint != nil {
+			if e.Checkpoint.Subject == subject && e.Checkpoint.FirstSeen > 0 {
+				if earliest == 0 || e.Checkpoint.FirstSeen < earliest {
+					earliest = e.Checkpoint.FirstSeen
+				}
+			}
+		}
+	}
+
+	return earliest
 }
