+++ social.go
@@ -37,11 +37,14 @@ type SocialEvent struct {
 	Signature string // Base64-encoded Ed25519 signature (optional)
 }
 
+// SignableContent returns the canonical string for signing (implements Signable)
+func (e *SocialEvent) SignableContent() string {
+	return fmt.Sprintf("%d:%s:%s:%s:%s:%s", e.Timestamp, e.Type, e.Actor, e.Target, e.Reason, e.Witness)
+}
+
 // Sign signs the social event with the given keypair
 func (e *SocialEvent) Sign(kp NaraKeypair) {
-	// Sign the content that makes up the ID (without the signature itself)
-	message := fmt.Sprintf("%d:%s:%s:%s:%s:%s", e.Timestamp, e.Type, e.Actor, e.Target, e.Reason, e.Witness)
-	e.Signature = kp.SignBase64([]byte(message))
+	e.Signature = SignContent(e, kp)
 }
 
 // Verify verifies the social event signature
@@ -49,8 +52,7 @@ func (e *SocialEvent) Verify(publicKey []byte) bool {
 	if e.Signature == "" {
 		return false
 	}
-	message := fmt.Sprintf("%d:%s:%s:%s:%s:%s", e.Timestamp, e.Type, e.Actor, e.Target, e.Reason, e.Witness)
-	return VerifySignatureBase64(publicKey, []byte(message), e.Signature)
+	return VerifyContent(e, publicKey, e.Signature)
 }
 
 // ComputeID generates a deterministic ID from event content
